Remove custom input
