+++ MANIFEST.in
@@ -13,7 +13,6 @@ recursive-include inginious/frontend/plugins/statistics *
 recursive-include inginious/frontend/plugins/utils *
 recursive-include inginious/frontend/plugins/multilang *
 recursive-include inginious/frontend/plugins/problem_bank *
-recursive-include inginious/frontend/plugins/custom_input *
 recursive-include inginious/frontend/plugins/UN_template *
 recursive-include inginious/frontend/plugins/UNCode *
 recursive-include inginious/frontend/plugins/register_students *
