+++ inginious/frontend/plugins/multilang/static/custom_input.js
